+++ batch/src/test/java/com/ImSnacks/NyeoreumnagiBatch/shortTermWeatherForecastTest/ApiCallerTest.java
@@ -1,4 +1,4 @@
-package com.ImSnacks.NyeoreumnagiBatch;
+package com.ImSnacks.NyeoreumnagiBatch.shortTermWeatherForecastTest;
 
 import jakarta.transaction.Transactional;
 import org.springframework.boot.test.context.SpringBootTest;
